[Evo][Timeline] Optimize external notifications
