+++ timeline/src/main/java/org/entcore/timeline/Timeline.java
@@ -20,19 +20,24 @@
 package org.entcore.timeline;
 
 import fr.wseduc.cron.CronTrigger;
-import fr.wseduc.webutils.Server;
 import fr.wseduc.webutils.request.filter.UserAuthFilter;
 import fr.wseduc.webutils.security.oauth.DefaultOAuthResourceProvider;
 
 import java.text.ParseException;
+import java.util.Map;
+import java.util.concurrent.ConcurrentMap;
 
 import org.entcore.common.http.BaseServer;
-import org.entcore.common.notification.TimelineMailer;
+import org.entcore.timeline.services.TimelineConfigService;
+import org.entcore.timeline.services.impl.DefaultTimelineConfigService;
+import org.entcore.timeline.services.impl.DefaultTimelineMailerService;
 import org.entcore.timeline.controllers.FlashMsgController;
 import org.entcore.timeline.controllers.TimelineController;
 import org.entcore.timeline.cron.DailyMailingCronTask;
 import org.entcore.timeline.cron.WeeklyMailingCronTask;
 import org.entcore.timeline.services.impl.FlashMsgRepositoryEventsSql;
+import org.vertx.java.core.impl.VertxInternal;
+import org.vertx.java.core.spi.cluster.ClusterManager;
 
 public class Timeline extends BaseServer {
 
@@ -42,19 +47,43 @@ public class Timeline extends BaseServer {
 		setOauthClientGrant(true);
 		addFilter(new UserAuthFilter(new DefaultOAuthResourceProvider(getEventBus(vertx))));
 		super.start();
-		addController(new TimelineController());
+
+		final Map<String, String> registeredNotifications;
+		Boolean cluster = (Boolean) vertx.sharedData().getMap("server").get("cluster");
+		if (Boolean.TRUE.equals(cluster)) {
+			ClusterManager cm = ((VertxInternal) vertx).clusterManager();
+			registeredNotifications = cm.getSyncMap("notificationsMap");
+		} else {
+			registeredNotifications = vertx.sharedData().getMap("notificationsMap");
+		}
+		final ConcurrentMap<String, String> eventsI18n = vertx.sharedData().getMap("timelineEventsI18n");
+
+		final TimelineConfigService configService = new DefaultTimelineConfigService("timeline.config");
+
+		final DefaultTimelineMailerService mailerService = new DefaultTimelineMailerService(vertx, container);
+		mailerService.setConfigService(configService);
+		mailerService.setRegisteredNotifications(registeredNotifications);
+		mailerService.setEventsI18n(eventsI18n);
+
+		final TimelineController timelineController = new TimelineController();
+		timelineController.setConfigService(configService);
+		timelineController.setMailerService(mailerService);
+		timelineController.setRegisteredNotifications(registeredNotifications);
+		timelineController.setEventsI18n(eventsI18n);
+		addController(timelineController);
+
 		addController(new FlashMsgController());
+
 		setRepositoryEvents(new FlashMsgRepositoryEventsSql());
 
-		TimelineMailer mailer = new TimelineMailer(vertx, Server.getEventBus(vertx), container, container.config().getInteger("users-loop-limit", 25));
 		final String dailyMailingCron = container.config().getString("daily-mailing-cron", "0 0 2 * * ?");
 		final String weeklyMailingCron = container.config().getString("weekly-mailing-cron", "0 0 5 ? * MON");
 		final int dailyDayDelta = container.config().getInteger("daily-day-delta", -1);
 		final int weeklyDayDelta = container.config().getInteger("weekly-day-delta", -1);
 
 		try {
-			new CronTrigger(vertx, dailyMailingCron).schedule(new DailyMailingCronTask(mailer, dailyDayDelta));
-			new CronTrigger(vertx, weeklyMailingCron).schedule(new WeeklyMailingCronTask(mailer, weeklyDayDelta));
+			new CronTrigger(vertx, dailyMailingCron).schedule(new DailyMailingCronTask(mailerService, dailyDayDelta));
+			new CronTrigger(vertx, weeklyMailingCron).schedule(new WeeklyMailingCronTask(mailerService, weeklyDayDelta));
 		} catch (ParseException e) {
 			log.error("Failed to start mailing crons.");
 		}
