+++ timeline/src/main/java/org/entcore/timeline/controllers/TimelineController.java
@@ -19,101 +19,70 @@
 
 package org.entcore.timeline.controllers;
 
-import java.io.IOException;
-import java.io.StringReader;
-import java.io.Writer;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.atomic.AtomicInteger;
-
 import fr.wseduc.bus.BusAddress;
 import fr.wseduc.rs.Delete;
 import fr.wseduc.rs.Get;
 import fr.wseduc.rs.Post;
 import fr.wseduc.rs.Put;
+import fr.wseduc.security.ActionType;
+import fr.wseduc.security.SecuredAction;
 import fr.wseduc.webutils.Either;
 import fr.wseduc.webutils.I18n;
 import fr.wseduc.webutils.Utils;
-import fr.wseduc.security.ActionType;
 import fr.wseduc.webutils.collections.TTLSet;
 import fr.wseduc.webutils.http.BaseController;
 import fr.wseduc.webutils.request.RequestUtils;
-
-import org.vertx.java.core.Handler;
-import org.vertx.java.core.Vertx;
-import org.vertx.java.core.VoidHandler;
-import org.vertx.java.core.eventbus.Message;
-import org.vertx.java.core.http.HttpServerRequest;
-import org.vertx.java.core.http.RouteMatcher;
-import org.vertx.java.core.impl.VertxInternal;
-import org.vertx.java.core.json.JsonArray;
-import org.vertx.java.core.json.JsonObject;
-import org.vertx.java.core.spi.cluster.ClusterManager;
-import org.vertx.java.platform.Container;
-
-import com.samskivert.mustache.Mustache;
-import com.samskivert.mustache.Template;
-import com.samskivert.mustache.Template.Fragment;
-
-import org.entcore.common.user.UserUtils;
 import org.entcore.common.http.filter.AdminFilter;
 import org.entcore.common.http.filter.AdmlOfStructures;
 import org.entcore.common.http.filter.ResourceFilter;
 import org.entcore.common.http.filter.SuperAdminFilter;
 import org.entcore.common.http.request.JsonHttpServerRequest;
 import org.entcore.common.notification.TimelineHelper;
-import org.entcore.common.notification.TimelineMailer;
 import org.entcore.common.notification.TimelineNotificationsLoader;
 import org.entcore.common.user.UserInfos;
-import fr.wseduc.security.SecuredAction;
+import org.entcore.common.user.UserUtils;
 import org.entcore.timeline.events.DefaultTimelineEventStore;
 import org.entcore.timeline.events.TimelineEventStore;
 import org.entcore.timeline.events.TimelineEventStore.AdminAction;
 import org.entcore.timeline.services.TimelineConfigService;
-import org.entcore.timeline.services.impl.DefaultTimelineConfigService;
+import org.entcore.timeline.services.TimelineMailerService;
+import org.vertx.java.core.Handler;
+import org.vertx.java.core.Vertx;
+import org.vertx.java.core.VoidHandler;
+import org.vertx.java.core.eventbus.Message;
+import org.vertx.java.core.http.HttpServerRequest;
+import org.vertx.java.core.http.RouteMatcher;
+import org.vertx.java.core.json.JsonArray;
+import org.vertx.java.core.json.JsonObject;
+import org.vertx.java.platform.Container;
 
-import static org.entcore.common.http.response.DefaultResponseHandler.*;
+import java.io.StringReader;
+import java.io.Writer;
+import java.util.*;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import static org.entcore.common.http.response.DefaultResponseHandler.arrayResponseHandler;
+import static org.entcore.common.http.response.DefaultResponseHandler.defaultResponseHandler;
 
 public class TimelineController extends BaseController {
 
 	private TimelineEventStore store;
 	private TimelineConfigService configService;
+	private TimelineMailerService mailerService;
+	private Map<String, String> registeredNotifications;
 	private ConcurrentMap<String, String> eventsI18n;
 	private HashMap<String, JsonObject> lazyEventsI18n = new HashMap<>();
-	private Map<String, String> registeredNotifications;
 	private Set<String> antiFlood;
 
 	//Declaring a TimelineHelper ensures the loading of the i18n/timeline folder.
 	private TimelineHelper timelineHelper;
 
-	private final String TIMELINE_CONFIG_COLLECTION = "timeline.config";
-
-	private TimelineMailer mailer;
-
 	public void init(Vertx vertx, Container container, RouteMatcher rm,
 			Map<String, fr.wseduc.webutils.security.SecuredAction> securedActions) {
 		super.init(vertx, container, rm, securedActions);
 		store = new DefaultTimelineEventStore();
-		configService = new DefaultTimelineConfigService(
-				TIMELINE_CONFIG_COLLECTION);
-		eventsI18n = vertx.sharedData().getMap("timelineEventsI18n");
 		timelineHelper = new TimelineHelper(vertx, eb, container);
-		Boolean cluster = (Boolean) vertx.sharedData().getMap("server")
-				.get("cluster");
-		if (Boolean.TRUE.equals(cluster)) {
-			ClusterManager cm = ((VertxInternal) vertx).clusterManager();
-			registeredNotifications = cm.getSyncMap("notificationsMap");
-		} else {
-			registeredNotifications = vertx.sharedData()
-					.getMap("notificationsMap");
-		}
-		mailer = new TimelineMailer(vertx, eb, container);
-
 		antiFlood = new TTLSet<>(container.config().getLong("antiFloodDelay", 3000l),
 				vertx, container.config().getLong("antiFloodClear", 3600 * 1000l));
 	}
@@ -122,68 +91,7 @@ public class TimelineController extends BaseController {
 	@Override
 	protected void setLambdaTemplateRequest(final HttpServerRequest request, final Map<String, Object> ctx) {
 		super.setLambdaTemplateRequest(request, ctx);
-
-		ctx.put("i18n", new Mustache.Lambda() {
-			@Override
-			public void execute(Template.Fragment frag, Writer out) throws IOException {
-				String key = frag.execute();
-				String language = Utils.getOrElse(I18n.acceptLanguage(request), "fr", false);
-
-				JsonObject timelineI18n;
-				if (!lazyEventsI18n.containsKey(language)) {
-					String i18n = eventsI18n.get(language.split(",")[0].split("-")[0]);
-					i18n = i18n != null ? i18n : "}";
-					timelineI18n = new JsonObject("{" + i18n.substring(0, i18n.length() - 1) + "}");
-					lazyEventsI18n.put(language, timelineI18n);
-				} else {
-					timelineI18n = lazyEventsI18n.get(language);
-				}
-
-				String translatedContents = I18n.getInstance().translate(key, getHost(request), language);
-				if (translatedContents.equals(key)) {
-					translatedContents = timelineI18n.getString(key, key);
-				}
-				Mustache.compiler().compile(translatedContents).execute(ctx, out);
-			}
-		});
-
-		ctx.put("host", new Mustache.Lambda() {
-			@Override
-			public void execute(Template.Fragment frag, Writer out) throws IOException{
-				String contents = frag.execute();
-				if(contents.matches("^(http://|https://).*")){
-					out.write(contents);
-				} else {
-					String host = getScheme(request) + "://" + getHost(request);
-					out.write(host + contents);
-				}
-			}
-		});
-
-		ctx.put("nested", new Mustache.Lambda() {
-			public void execute(Fragment frag, Writer out) throws IOException {
-				String nestedTemplateName = frag.execute();
-				String nestedTemplate = (String) ctx.get(nestedTemplateName);
-				if(nestedTemplate != null)
-					Mustache.compiler().compile(nestedTemplate).execute(ctx, out);
-			}
-		});
-
-		ctx.put("nestedArray", new Mustache.Lambda() {
-			public void execute(Fragment frag, Writer out) throws IOException {
-				String nestedTemplatePos = frag.execute();
-				JsonArray nestedArray = new JsonArray((List<Object>) ctx.get("nestedTemplatesArray"));
-				try {
-					JsonObject nestedTemplate = (JsonObject) nestedArray.get(Integer.parseInt(nestedTemplatePos) - 1);
-					ctx.putAll(nestedTemplate.getObject("params", new JsonObject()).toMap());
-					Mustache.compiler()
-						.compile(nestedTemplate.getString("template", ""))
-						.execute(ctx, out);
-				} catch(NumberFormatException e) {
-					log.error("Mustache compiler error while parsing a nested template array lambda.");
-				}
-			}
-		});
+		TimelineLambda.setLambdaTemplateRequest(request, ctx, eventsI18n, lazyEventsI18n);
 	}
 
 	@Get("/timeline")
@@ -422,14 +330,14 @@ public class TimelineController extends BaseController {
 	@SecuredAction(type = ActionType.RESOURCE, value = "")
 	@ResourceFilter(SuperAdminFilter.class)
 	public void performDailyMailing(final HttpServerRequest request) {
-		mailer.sendDailyMails(0, defaultResponseHandler(request));
+		mailerService.sendDailyMails(0, defaultResponseHandler(request));
 	}
 
 	@Get("/performWeeklyMailing")
 	@SecuredAction(type = ActionType.RESOURCE, value = "")
 	@ResourceFilter(SuperAdminFilter.class)
 	public void performWeeklyMailing(final HttpServerRequest request) {
-		mailer.sendWeeklyMails(0, defaultResponseHandler(request));
+		mailerService.sendWeeklyMails(0, defaultResponseHandler(request));
 	}
 
 	@Get("/allowLanguages")
@@ -695,6 +603,11 @@ public class TimelineController extends BaseController {
 			if (sender == null || sender.startsWith("no-reply") || antiFlood.add(sender)) {
 				store.add(json, new Handler<JsonObject>() {
 					public void handle(JsonObject result) {
+						mailerService.sendImmediateMails(
+								new JsonHttpServerRequest(json.getObject("request")),
+								json.getString("notificationName"), json.getObject("notification"), json.getObject("params"),
+								json.getArray("recipientsIds")
+						);
 						handler.handle(result);
 					}
 				});
@@ -723,104 +636,6 @@ public class TimelineController extends BaseController {
 							.putArray("types", types));
 				}
 			});
-			break;
-		case "get-notification-properties":
-			getNotificationProperties(json.getString("key"),
-					new Handler<Either<String, JsonObject>>() {
-						public void handle(Either<String, JsonObject> event) {
-							if (event.isLeft()) {
-								message.reply(new JsonObject()
-										.putString("status", "error")
-										.putString("message",
-												event.left().getValue()));
-							} else {
-								message.reply(new JsonObject()
-										.putString("status", "ok")
-										.putObject("result",
-												event.right().getValue()));
-							}
-						}
-					});
-			break;
-		case "list-notifications-defaults":
-			configService.list(new Handler<Either<String,JsonArray>>() {
-				public void handle(Either<String, JsonArray> event) {
-					if (event.isLeft()) {
-						message.reply(new JsonObject()
-								.putString("status", "error")
-								.putString("message",
-										event.left().getValue()));
-					} else {
-						JsonArray config = event.right().getValue();
-
-						JsonArray notificationsList = new JsonArray();
-						for (String key : registeredNotifications.keySet()) {
-							JsonObject notif = new JsonObject(registeredNotifications.get(key));
-							notif.putString("key", key);
-							for(Object notifConfigObj: config){
-								JsonObject notifConfig = (JsonObject) notifConfigObj;
-								if (notifConfig.getString("key", "").equals(key)) {
-									notif.putString("defaultFrequency",
-											notifConfig.getString("defaultFrequency", notif.getString("defaultFrequency")));
-									notif.putString("restriction",
-											notifConfig.getString("restriction", notif.getString("restriction")));
-									break;
-								}
-							}
-							notificationsList.add(notif);
-						}
-						message.reply(new JsonObject()
-								.putString("status", "ok")
-								.putArray("results", notificationsList));
-					}
-				}
-			});
-			break;
-		case "process-timeline-template":
-			final HttpServerRequest request = new JsonHttpServerRequest(
-					message.body().getObject("request", new JsonObject()));
-			final JsonObject parameters = message.body().getObject("parameters", new JsonObject());
-			final String resourceName = message.body().getString("resourceName", "");
-			if(message.body().getBoolean("reader", false)){
-				final StringReader templateReader = new StringReader(message.body().getString("template"));
-				processTemplate(request, parameters, resourceName, templateReader, new Handler<Writer>() {
-					public void handle(Writer writer) {
-						message.reply(
-							new JsonObject()
-								.putString("status", "ok")
-								.putString("processedTemplate", writer.toString()));
-					}
-				});
-
-			} else {
-				processTemplate(request, message.body().getString("template", ""), parameters, new Handler<String>() {
-					public void handle(String template) {
-						message.reply(
-							new JsonObject()
-							 	.putString("status", "ok")
-								.putString("processedTemplate", template));
-					}
-				});
-			}
-
-			break;
-		case "translate-timeline":
-			final JsonArray i18nKeys = message.body().getArray("i18nKeys", new JsonArray());
-			final String language = message.body().getString("language", "fr");
-			final String domain = message.body().getString("domain", I18n.DEFAULT_DOMAIN);
-			String i18n = eventsI18n
-				.get(language.split(",")[0].split("-")[0]);
-			JsonObject timelineI18n = new JsonObject(
-					"{" + i18n.substring(0, i18n.length() - 1) + "}");
-			timelineI18n.mergeIn(I18n.getInstance().load(language, domain));
-			JsonArray translations = new JsonArray();
-			for(Object keyObj : i18nKeys){
-				String key = (String) keyObj;
-				translations.add(timelineI18n.getString(key, key));
-			}
-			message.reply(new JsonObject()
-				.putString("status", "ok")
-				.putArray("translations", translations));
 			break;
 		default:
 			message.reply(new JsonObject().putString("status", "error")
@@ -860,38 +675,19 @@ public class TimelineController extends BaseController {
 		});
 	}
 
-	private void getNotificationProperties(final String notificationKey,
-			final Handler<Either<String, JsonObject>> handler) {
-		configService.list(new Handler<Either<String, JsonArray>>() {
-			public void handle(Either<String, JsonArray> event) {
-				if (event.isLeft()) {
-					handler.handle(new Either.Left<String, JsonObject>(
-							event.left().getValue()));
-					return;
-				}
-				final String notificationStr = registeredNotifications
-						.get(notificationKey.toLowerCase());
-				if (notificationStr == null) {
-					handler.handle(new Either.Left<String, JsonObject>(
-							"invalid.notification.key"));
-					return;
-				}
-				final JsonObject notification = new JsonObject(notificationStr);
-				for (Object notifConfigObj : event.right().getValue()) {
-					JsonObject notifConfig = (JsonObject) notifConfigObj;
-					if (notifConfig.getString("key", "")
-							.equals(notificationKey.toLowerCase())) {
-						notification.putString("defaultFrequency",
-								notifConfig.getString("defaultFrequency", ""));
-						notification.putString("restriction",
-								notifConfig.getString("restriction", ""));
-						break;
-					}
-				}
-				handler.handle(
-						new Either.Right<String, JsonObject>(notification));
-			}
-		});
+	public void setConfigService(TimelineConfigService configService) {
+		this.configService = configService;
 	}
 
+	public void setMailerService(TimelineMailerService mailerService) {
+		this.mailerService = mailerService;
+	}
+
+	public void setRegisteredNotifications(Map<String, String> registeredNotifications) {
+		this.registeredNotifications = registeredNotifications;
+	}
+
+	public void setEventsI18n(ConcurrentMap<String, String> eventsI18n) {
+		this.eventsI18n = eventsI18n;
+	}
 }
