+++ timeline/src/main/java/org/entcore/timeline/cron/DailyMailingCronTask.java
@@ -19,7 +19,7 @@
 
 package org.entcore.timeline.cron;
 
-import org.entcore.common.notification.TimelineMailer;
+import org.entcore.timeline.services.TimelineMailerService;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.json.JsonObject;
 import org.vertx.java.core.logging.Logger;
@@ -30,18 +30,18 @@ import fr.wseduc.webutils.Either;
 public class DailyMailingCronTask implements Handler<Long> {
 
 	private static final Logger log = LoggerFactory.getLogger(DailyMailingCronTask.class);
-	private final TimelineMailer mailer;
+	private final TimelineMailerService mailerService;
 	private final int dayDelta;
 
-	public DailyMailingCronTask(TimelineMailer mailer, int dayDelta){
-		this.mailer = mailer;
+	public DailyMailingCronTask(TimelineMailerService mailerService, int dayDelta){
+		this.mailerService = mailerService;
 		this.dayDelta = dayDelta;
 	}
 
 	@Override
 	public void handle(Long event) {
 		log.info("[Daily mailing] Starting ...");
-		mailer.sendDailyMails(dayDelta, new Handler<Either<String,JsonObject>>() {
+		mailerService.sendDailyMails(dayDelta, new Handler<Either<String,JsonObject>>() {
 			public void handle(Either<String, JsonObject> event) {
 				if(event.isLeft()){
 					log.error("[Daily mailing] Error encountered : " + event.left().getValue());
