+++ timeline/src/main/java/org/entcore/timeline/cron/WeeklyMailingCronTask.java
@@ -19,7 +19,7 @@
 
 package org.entcore.timeline.cron;
 
-import org.entcore.common.notification.TimelineMailer;
+import org.entcore.timeline.services.TimelineMailerService;
 import org.vertx.java.core.Handler;
 import org.vertx.java.core.json.JsonObject;
 import org.vertx.java.core.logging.Logger;
@@ -30,18 +30,18 @@ import fr.wseduc.webutils.Either;
 public class WeeklyMailingCronTask implements Handler<Long> {
 
 	private static final Logger log = LoggerFactory.getLogger(WeeklyMailingCronTask.class);
-	private final TimelineMailer mailer;
+	private final TimelineMailerService mailerService;
 	private final int dayDelta;
 
-	public WeeklyMailingCronTask(TimelineMailer mailer, int dayDelta){
-		this.mailer = mailer;
+	public WeeklyMailingCronTask(TimelineMailerService mailer, int dayDelta){
+		this.mailerService = mailer;
 		this.dayDelta = dayDelta;
 	}
 
 	@Override
 	public void handle(Long event) {
 		log.info("[Weekly mailing] Starting ...");
-		mailer.sendWeeklyMails(dayDelta, new Handler<Either<String,JsonObject>>() {
+		mailerService.sendWeeklyMails(dayDelta, new Handler<Either<String,JsonObject>>() {
 			public void handle(Either<String, JsonObject> event) {
 				if(event.isLeft()){
 					log.error("[Weekly mailing] Error encountered : " + event.left().getValue());
