+++ timeline/src/main/java/org/entcore/timeline/services/TimelineConfigService.java
@@ -27,7 +27,6 @@ import org.vertx.java.core.json.JsonObject;
 
 public interface TimelineConfigService{
 
-	public void create(JsonObject data, UserInfos user, Handler<Either<String, JsonObject>> handler);
 	public void upsert(JsonObject data, Handler<Either<String, JsonObject>> handler);
 	public void list(Handler<Either<String, JsonArray>> handler);
